Migrate to ESLint Flat Config & [Snyk] Security upgrade eslint from 8.56.0 to 8.57.0 (#210)
* fix: package.json to reduce vulnerabilities

The following vulnerabilities are fixed with an upgrade:
- https://snyk.io/vuln/SNYK-JS-INFLIGHT-6095116

* update typescript-eslint to v7
- `@typescript-eslint/recommended` is included in `@typescript-eslint/strict`
- `svelte/all` helps us
- almost of all browsers can parse es2023

* audit fix

* remove VSCode tabSize setting to take the advantage of tab indentation

* don't use ESLint `--cache` option

Typed lint rules almost always have dependencies on types across files in practice. https://typescript-eslint.io/troubleshooting/#can-i-use-eslints---cache-with-typescript-eslint

* update linter, typechecker, and formatter

* migrate: to ESLint Flat Config
- https://eslint.org/docs/v8.x/use/configure/migration-guide
- https://typescript-eslint.io/blog/announcing-typescript-eslint-v7/#new-features---flat-config-support
- https://sveltejs.github.io/eslint-plugin-svelte/user-guide/#new-config-eslint-config-js
- https://zenn.dev/babel/articles/eslint-flat-config-for-babel#flat-config-への移行-tips

* add ESLint option reporting unused disable directives

---------
