+++ package.json
@@ -37,31 +37,30 @@
 		"test": "vitest",
 		"vitest:all": "run-s typecheck:vitest test",
 		"format:dry": "prettier --cache --check .",
-		"lint:dry": "eslint --cache --cache-location ./node_modules/.cache/eslint/.eslintcache .",
+		"lint:dry": "eslint .",
 		"format": "prettier --cache --write .",
-		"lint": "eslint --fix --cache --cache-location ./node_modules/.cache/eslint/.eslintcache ."
+		"lint": "eslint --fix ."
 	},
 	"devDependencies": {
 		"@sveltejs/adapter-vercel": "^5.1.1",
 		"@sveltejs/kit": "^2.5.4",
 		"@sveltejs/vite-plugin-svelte": "^3.0.2",
-		"@typescript-eslint/eslint-plugin": "^6.21.0",
-		"@typescript-eslint/parser": "^6.21.0",
-		"@typescript/lib-dom": "npm:@types/web@^0.0.130",
+		"@typescript/lib-dom": "npm:@types/web@^0.0.142",
 		"@vitest/coverage-v8": "^1.4.0",
-		"better-typescript-lib": "^2.6.0",
-		"eslint": "^8.56.0",
+		"better-typescript-lib": "~2.7.0",
+		"eslint": "^8.57.0",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.35.1",
+		"eslint-plugin-svelte": "^2.37.0",
 		"npm-run-all": "^4.1.5",
 		"prettier": "^3.2.5",
-		"prettier-plugin-svelte": "^3.2.2",
+		"prettier-plugin-svelte": "^3.2.3",
 		"sass": "^1.72.0",
 		"svelte": "^4.2.12",
-		"svelte-check": "^3.6.7",
+		"svelte-check": "^3.6.9",
 		"svelte-scrollto": "0.2.0",
 		"tslib": "^2.6.2",
 		"typescript": "~5.4.2",
+		"typescript-eslint": "^7.6.0",
 		"typescript-eslint-language-service": "^5.0.5",
 		"vite": "^5.1.7",
 		"vitest": "^1.4.0",
