+++ src/lib/games/AppBanner.svelte
@@ -2,8 +2,10 @@
 import { onMount } from 'svelte';
 let dialogElement: HTMLDialogElement;
 onMount(() => {
+	// eslint-disable-next-line @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-member-access
 	dialogElement.showModal();
 	setTimeout(() => {
+		// eslint-disable-next-line @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-member-access
 		dialogElement.close();
 	}, 2400);
 });
