+++ src/lib/games/quantum-tictactoe/QuantumTTT.ts
@@ -70,8 +70,8 @@ export default class QuantumTTT {
 		return this.state.currentSubTurn < 2 ? 'Y' : 'X';
 	}
 
-	timer(): void {
-		if (this.whoseTurn() === 'X') {
+	timer = (): void => {
+		if (this.whoseTurn() === 'X' && !this.state.cycleSquares?.length) {
 			if (this.state.leftTimes.X <= 0) {
 				this.setState({
 					isOver: true,
@@ -82,7 +82,7 @@ export default class QuantumTTT {
 			}
 		}
 
-		if (this.whoseTurn() === 'Y') {
+		if (this.whoseTurn() === 'Y' && !this.state.cycleSquares?.length) {
 			if (this.state.leftTimes.Y <= 0) {
 				this.setState({
 					isOver: true,
@@ -92,7 +92,7 @@ export default class QuantumTTT {
 				this.setLeftTimes({ Y: this.state.leftTimes.Y - 1 });
 			}
 		}
-	}
+	};
 
 	// dispatches click to appropriate handler based on state
 	handleSquareClick(i: SquareType): StatusType {
@@ -244,7 +244,6 @@ function _calculateWinners(squares: Readonly<ConstArray<MarkType | null, 9>>): W
 	const winners: WinnersType = [];
 	for (const line of lines) {
 		const [s1, s2, s3] = [squares[line[0]], squares[line[1]], squares[line[2]]];
-		// eslint-disable-next-line @typescript-eslint/prefer-string-starts-ends-with
 		if (s1 && s2 && s3 && s1[0] === s2[0] && s1[0] === s3[0]) {
 			const subscripts = [s1[1], s2[1], s3[1]].map(Number) as ConstArray<TurnNumType, 3>;
 			winners.push([Math.max(...subscripts) as TurnNumType, s1[0] as PlayerType, line]);
