+++ src/routes/+error.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
 import { page } from '$app/stores';
+$: ({ status, error } = $page);
 </script>
 
 <svelte:head>
@@ -9,6 +10,6 @@ import { page } from '$app/stores';
 	<meta property="og:title" content="Quantum Game Arena" />
 </svelte:head>
 
-<main style="padding: var(--header-height) 1rem 0;">
-	<h1>{$page.status}: {$page.error?.message}</h1>
+<main style:padding="var(--header-height) 1rem 0">
+	<h1>{status}: {error?.message}</h1>
 </main>
