+++ tsconfig.eslint.json
@@ -1,8 +1,8 @@
 {
 	"extends": "./tsconfig.json",
 	"include": [
-		"./.svelte-kit/ambient.d.ts",
-		"./.svelte-kit/non-ambient.d.ts",
+		".svelte-kit/ambient.d.ts",
+		".svelte-kit/non-ambient.d.ts",
 		"./.svelte-kit/types/**/$types.d.ts",
 		"./svelte.config.js",
 		"./vite.config.ts",
@@ -11,7 +11,7 @@
 		"./tests/**/*.ts",
 		"./tests/**/*.svelte",
 
-		"./.eslintrc.cjs"
+		"./eslint.config.js"
 	],
 	"exclude": ["./node_modules/**"]
 }
